Show time slider widget, animate with filter
Instead of changing the underlying featureCollection with each
animation step, use a filter on the timestamp property. I was hoping
that this makes the animation more performant - it looks a little bit
like it does, but I cannot confirm for sure.

The animation runs once and then stops. Nicer would be a play/pause
button.

With this there is no panning during the animation of the map anymore,
but I thought having a time slider that also shows the displayed time
range is still better.

Function naming and code organization makes little sense, feel free to
refactor, or revert to the previous animation approach if you want.